add: confirm delete appointment

+++ src/app/dashboard/components/today/today.component.ts
@@ -11,6 +11,7 @@ import { heroChevronLeft, heroChevronRight, heroTrash } from '@ng-icons/heroicon
 import { RouterLink } from '@angular/router';
 import { NgClass } from '@angular/common';
 import { AppointmentResponse } from '@app/features/appointments/models/responses/appointments.response';
+import { ConfirmDialogComponent } from '@app/shared/components/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-today',
@@ -95,16 +96,21 @@ export class TodayComponent {
   }
 
   deleteAppointment(appointment: AppointmentResponse) {
-    this.appointmentsService.signalAppointmentSelected.set(appointment);
-    this.appointmentsService.deleteAppointment().subscribe(() => {
-      const currentPageIndex = this.currentPage(); 
-      const turnosRestantes = this.paginatedAppointments();
-      if (turnosRestantes.length === 0 && currentPageIndex > 0) {
-        this.currentPage.set(currentPageIndex - 1);
-      }
-    });
+    this.modalService.openWithResult(ConfirmDialogComponent, {}, {message: '¿Estás seguro de que querés eliminar este cliente?'}).subscribe((confirmed: boolean) => {
+        if (confirmed) {
+          this.appointmentsService.signalAppointmentSelected.set(appointment);
+          this.appointmentsService.deleteAppointment().subscribe(() => {
+            const currentPageIndex = this.currentPage();
+            const turnosRestantes = this.paginatedAppointments();
+            if (turnosRestantes.length === 0 && currentPageIndex > 0) {
+              this.currentPage.set(currentPageIndex - 1);
+            }
+          });
+          this.alertService.showSuccess('Turno eliminado');
+        }
+      });
   }
-  
+
   returnIfPlaceForNewAppointments() {
     const appointments = this.signalAppointmentsForDate();
     if (appointments.length === 0) return true;
